# use read lock to avoid io hang during heartbeat