Io hanged during heartbeat because of the write lock. Hanging time depends on volume number. Since deleting expired volume is low frequency，we can do this out of volume collecting.

Here is a simple comparing test with 500 volumes in a volume server:
test command: 
weed benchmark -master=localhost:9334 -c=32 -n=50000 -size=81920 -read=false 

before optimization:

Completed 25454 of 50000 requests, 50.9% 1274.0/s 99.6MB/s
Completed 26714 of 50000 requests, 53.4% 1260.0/s 98.5MB/s
Completed 27828 of 50000 requests, 55.7% 1114.0/s 87.1MB/s
Completed 28963 of 50000 requests, 57.9% 1135.0/s 88.7MB/s
Completed 30233 of 50000 requests, 60.5% 1270.0/s 99.3MB/s
Completed 31507 of 50000 requests, 63.0% 1274.0/s 99.6MB/s
Completed 32732 of 50000 requests, 65.5% 1225.0/s 95.7MB/s
Completed 33847 of 50000 requests, 67.7% 1115.0/s 87.1MB/s
Completed 34963 of 50000 requests, 69.9% 1116.0/s 87.2MB/s
Completed 36222 of 50000 requests, 72.4% 1259.0/s 98.4MB/s
Completed 37491 of 50000 requests, 75.0% 1269.0/s 99.2MB/s
Completed 38725 of 50000 requests, 77.5% 1234.0/s 96.4MB/s
Completed 39827 of 50000 requests, 79.7% 1102.0/s 86.1MB/s
Completed 40957 of 50000 requests, 81.9% 1130.0/s 88.3MB/s
Completed 42220 of 50000 requests, 84.4% 1263.0/s 98.7MB/s
Completed 43500 of 50000 requests, 87.0% 1280.0/s 100.0MB/s
Completed 44738 of 50000 requests, 89.5% 1238.0/s 96.8MB/s
Completed 45852 of 50000 requests, 91.7% 1114.0/s 87.1MB/s
Completed 47007 of 50000 requests, 94.0% 1155.0/s 90.3MB/s
Completed 48251 of 50000 requests, 96.5% 1244.0/s 97.2MB/s
Completed 49515 of 50000 requests, 99.0% 1264.0/s 98.8MB/s

Concurrency Level:      32
Time taken for tests:   41.368 seconds
Complete requests:      50000
Failed requests:        0
Total transferred:      4097570372 bytes
Requests per second:    1208.67 [#/sec]
Transfer rate:          96730.40 [Kbytes/sec]

Connection Times (ms)
              min      avg        max      std
Total:        3.1      26.4       260.7      16.0



after optimization:
Completed 25861 of 50000 requests, 51.7% 1272.0/s 99.4MB/s
Completed 27103 of 50000 requests, 54.2% 1242.0/s 97.1MB/s
Completed 28312 of 50000 requests, 56.6% 1209.0/s 94.5MB/s
Completed 29555 of 50000 requests, 59.1% 1243.0/s 97.1MB/s
Completed 30805 of 50000 requests, 61.6% 1250.0/s 97.7MB/s
Completed 32066 of 50000 requests, 64.1% 1261.0/s 98.6MB/s
Completed 33275 of 50000 requests, 66.5% 1209.0/s 94.5MB/s
Completed 34476 of 50000 requests, 69.0% 1201.0/s 93.9MB/s
Completed 35727 of 50000 requests, 71.5% 1251.0/s 97.8MB/s
Completed 36974 of 50000 requests, 73.9% 1247.0/s 97.5MB/s
Completed 38235 of 50000 requests, 76.5% 1261.0/s 98.6MB/s
Completed 39485 of 50000 requests, 79.0% 1250.0/s 97.7MB/s
Completed 40680 of 50000 requests, 81.4% 1195.0/s 93.4MB/s
Completed 41935 of 50000 requests, 83.9% 1255.0/s 98.1MB/s
Completed 43176 of 50000 requests, 86.4% 1241.1/s 97.0MB/s
Completed 44409 of 50000 requests, 88.8% 1233.0/s 96.4MB/s
Completed 45653 of 50000 requests, 91.3% 1244.0/s 97.2MB/s
Completed 46865 of 50000 requests, 93.7% 1212.0/s 94.7MB/s
Completed 48118 of 50000 requests, 96.2% 1253.0/s 97.9MB/s
Completed 49367 of 50000 requests, 98.7% 1249.0/s 97.6MB/s

Concurrency Level:      32
Time taken for tests:   40.491 seconds
Complete requests:      50000
Failed requests:        0
Total transferred:      4097578598 bytes
Requests per second:    1234.84 [#/sec]
Transfer rate:          98824.97 [Kbytes/sec]

Connection Times (ms)
              min      avg        max      std
Total:        3.4      25.9       186.4      6.3